requirements.txt
aiogram==3.10.0
fastapi==0.112.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
websockets==12.0
pillow==10.4.0
protobuf==5.27.3